Allow giving the container permission to access the host's /dev
This is basically rendering the container completely useless in terms of
isolation - whetever is running in the container has full *root access*
to the host.
Therefore this option as well as all other options giving the container
extra abailities are locked behing a config switch as well as
configuration file change.
Unfortunately some build processes for images need the additional
capabilities, so we have to allow this for specialized use cases.

There is a chance that we might actually replace this functionality with
more fine-grained /dev access control in the future.