Allow turning on d/rules clean on the host, disable it by default
Previously, d/rules clean would be run on the host, without all build
dependencies installed. This can lead to puzzling issues when working
with some packages.
This change disables running of d/rules clean on the host entirely.
Instead, the source tree is cleaned in the container again before being
built. At that point, all dependencies required by the package are also
installed.
To return to the previous behavior (but with dependency checks
re-enabled), users can pass the --clean-source flag to debspawn.
This resolves #6