Ensure all data lands in its intended directories when installing
This is probably misusing the setuptools system, but I found no other
way to deal with it. This is horrendously complicated for the simple
tasks it is set out to achieve (installing Python packages) with so many
different ways to achieve the same thing, while simultaneously not being
able to do simple system-wide installations on Linux.
So, we work around this for now, until I actually find a proper
solution.