Use dpkg --print-architecture to determine arch
Basically, the only thing that dpkg-architecture is being used for is
`dpkg-architecture -qDEB_HOST_ARCH`. Its output is used as a fallback
when no architecture is provided by the user. It is intended to serve
the architecture of the running system (even though that's not quite
what it does). A simpler way to do that is dpkg --print-architecture and
in doing so, the dpkg-dev dependency is avoided.