Rework how external system files are installed
Pip & co. really don't want us to install into system directories, so
the only long-term choice is likely to provide an additional script for
this, which this commit does.
That way, people can install via pip and for additional support run the
install-sysdata.py script.
If setup.py is used directly, we'll make an attempt to install the data
files in the right places as well.