Allow containers that run a custom command to be booted as well
This is a pretty complex setup, but is quite handy for some automation
tasks. Unfortunately the container doesn't seem to communicate when it
is booted up properly, so we have a nasty timeout in the code for now
that needs to be resolved properly as soon as a solution becomes
available.