Use useradd instead of adduser
The adduser tool was used from Debian's default set of tools on a
minimal buildd installation, so we need to use lower-level tools here.
Depending on the system we do get a UID range warning now, so
potentially need to address that issue in future by raising UID_MAX to
something higher than 6000.