Work around nspawn wanting to ensure its user-home exists
Since we can not prevent nspawn from creating the directory, and the
nspawn --inaccessible flag requires a permanently existing directory in
the container (which is undesirable), we move the builder's home to a
temporary location and then remove all permissions from the directory.
It is not a foolproof method, but works very well in tests. In order to
fail, it would require package builds to actively want to break the
system and want to write into HOME (by ignoring the HOME env var, and
changing permissions on the extracted HOME directory, and then write to
it). This is a very unlikely scenario.
So, until nspawn grows an option to not create a home dir for users it
switches to, this will be an acceptable workaround.

Resolves: #32